refactor: move `entity-kind` from `data-model` to `runner`.
`data-model` never used this module: every consumer was external, and the only
in-package references were its own export entry. It describes the URI scheme
that wraps a tagged hash -- by its own account, the part that lives OUTSIDE the
tagged-hash string -- while `data-model` owns the hash itself. The one thing
that ever tied it to `data-model` was the `fid2:computed:<hash>` tag format,
retired in #4659 before it shipped.

`runner` is where the concept already lives. It is the only package that mints
(`uriSchemeForEntityKind`); everyone else reads. Its collaborator `uri-utils.ts`
-- `toURI` consumes `uriSchemeForEntityKind`, `fromURI` consumes
`entityUriSchemePrefix` -- now sits beside it rather than across a package
boundary neither side could see over.

Exported as `@commonfabric/runner/entity-kind`, mirroring `./slugs`: a narrow
subpath, so `shell` and `ui` (which need three pure functions between them) keep
importing a leaf rather than the runtime. No cycle is created -- `data-model`
has no need of it, so no `data-model` -> `runner` edge appears.

`EntityKind` is deliberately not renamed here, though it collides with an
unrelated `EntityKind` in `state-inspector/model.ts`.

Consumers re-pointed: 8 in `runner` (now relative), 4 in `shell`, 1 in `ui`.
No behavior change.
