+++ packages/runner/src/storage/transaction-summary.ts
@@ -5,7 +5,7 @@
  * into concise summaries suitable for LLMs to help humans debug software behavior.
  */
 
-import { entityUriSchemePrefix } from "@commonfabric/data-model/entity-kind";
+import { entityUriSchemePrefix } from "../entity-kind.ts";
 import { toCompactDebugString } from "@commonfabric/data-model/value-debug";
 import type {
   IExtendedStorageTransaction,
