+++ packages/runner/src/uri-utils.ts
@@ -3,7 +3,7 @@ import {
   entityUriSchemePrefix,
   hasEntityUriScheme,
   uriSchemeForEntityKind,
-} from "@commonfabric/data-model/entity-kind";
+} from "./entity-kind.ts";
 import { FabricHash } from "@commonfabric/data-model/fabric-primitives";
 import { hashOf } from "@commonfabric/data-model/value-hash";
 import { toCompactDebugString } from "@commonfabric/data-model/value-debug";
