+++ packages/ui/src/v2/utils/mention-id.ts
@@ -1,4 +1,4 @@
-import { entityUriSchemePrefix } from "@commonfabric/data-model/entity-kind";
+import { entityUriSchemePrefix } from "@commonfabric/runner/entity-kind";
 
 /**
  * The id embedded in wiki-link mention text (`[[Name (<id>)]]`) is the BARE
